+++ projects/CoinFlip/CoinFlip.Model/src/Market/Offer.cs
@@ -8,7 +8,7 @@ namespace CoinFlip
 	[DebuggerDisplay("{Description,nq}")]
 	public struct Offer : IComparable<Offer>
 	{
-		public Offer(Unit<double> price_q2b, Unit<double> amount_base)
+		public Offer(Unit<decimal> price_q2b, Unit<decimal> amount_base)
 		{
 			Debug.Assert(!price_q2b.IsUnit(0));
 			PriceQ2B = price_q2b;
@@ -16,19 +16,19 @@ namespace CoinFlip
 		}
 
 		/// <summary>The price (to buy or sell) (in Quote/Base)</summary>
-		public Unit<double> PriceQ2B { get; set; }
+		public Unit<decimal> PriceQ2B { get; set; }
 
 		/// <summary>The volume (in base currency)</summary>
-		public Unit<double> AmountBase { get; set; }
+		public Unit<decimal> AmountBase { get; set; }
 
 		/// <summary>The volume (in quote currency)</summary>
-		public Unit<double> AmountQuote => AmountBase * PriceQ2B;
+		public Unit<decimal> AmountQuote => AmountBase * PriceQ2B;
 
 		/// <summary>The 'In' side of the offer</summary>
-		public Unit<double> AmountIn(ETradeType tt) => tt.AmountIn(AmountBase, PriceQ2B);
+		public Unit<decimal> AmountIn(ETradeType tt) => tt.AmountIn(AmountBase, PriceQ2B);
 
 		/// <summary>The 'Out' side of the offer</summary>
-		public Unit<double> AmountOut(ETradeType tt) => tt.AmountOut(AmountBase, PriceQ2B);
+		public Unit<decimal> AmountOut(ETradeType tt) => tt.AmountOut(AmountBase, PriceQ2B);
 
 		/// <summary>Orders are compared by price</summary>
 		public int CompareTo(Offer rhs) => PriceQ2B.CompareTo(rhs.PriceQ2B);
@@ -39,7 +39,7 @@ namespace CoinFlip
 		/// <summary>Check this order against the limits given in 'pair'</summary>
 		public Exception Validate(TradePair pair)
 		{
-			if (PriceQ2B <= 0.0)
+			if (PriceQ2B <= 0m)
 				return new Exception($"Offer price ({PriceQ2B.ToString(6)}) is <= 0");
 
 			if (!pair.AmountRangeBase.Contains(AmountBase))
