+++ projects/CoinFlip/ExchApi.Binance/src/WebSocket/MarketDataCache.cs
@@ -293,16 +293,16 @@ namespace Binance.API
 			{
 				return offers.Select(x => new Offer
 				{
-					Price = double.Parse(x[0], CultureInfo.InvariantCulture),
-					Quantity = double.Parse(x[1], CultureInfo.InvariantCulture),
+					Price = decimal.Parse(x[0], CultureInfo.InvariantCulture),
+					Quantity = decimal.Parse(x[1], CultureInfo.InvariantCulture),
 				}).ToList();
 			}
 
 			/// </summary>
 			public class Offer
 			{
-				public double Price { get; set; }
-				public double Quantity { get; set; }
+				public decimal Price { get; set; }
+				public decimal Quantity { get; set; }
 			}
 
 			private class DepthDeltaArrayConverter :JsonConverter
@@ -313,8 +313,8 @@ namespace Binance.API
 					var list = new List<Offer>();
 					foreach (var tradePrice in JArray.Load(reader))
 					{
-						var price = tradePrice.ElementAt(0).ToObject<double>();
-						var quantity = tradePrice.ElementAt(1).ToObject<double>();
+						var price = tradePrice.ElementAt(0).ToObject<decimal>();
+						var quantity = tradePrice.ElementAt(1).ToObject<decimal>();
 						list.Add(new Offer { Price = price, Quantity = quantity });
 					}
 					return list;
